- Fixed typing issue

+++ LemonadeStand/lib/Ingredients.ml
@@ -1,5 +1,5 @@
 type price = {
-  amt : int;
+  amt : float;
   total_cost : float;
   unit_price : float;
 }
@@ -17,9 +17,9 @@ type bounds = {
 }
 
 type cup_contains = {
-  lemon : int;
-  sugar : int;
-  water : int;
+  lemon : float;
+  sugar : float;
+  water : float;
 }
 
 let set_bounds_lemons = { upper_bound = 1.; lower_bound = 2.; optimal = 1.5 }
@@ -28,10 +28,10 @@ let set_bounds_water = { upper_bound = 0.75; lower_bound = 1.25; optimal = 1. }
 
 let purchase_options =
   {
-    lemon = { amt = 10; total_cost = 2.0; unit_price = 10.0 /. 2.0 };
-    cup = { amt = 10; total_cost = 2.0; unit_price = 10.0 /. 2.0 };
+    lemon = { amt = 10.; total_cost = 2.0; unit_price = 10.0 /. 2.0 };
+    cup = { amt = 10.; total_cost = 2.0; unit_price = 10.0 /. 2.0 };
     sugar =
-      { amt = 10; total_cost = 2.0; unit_price = 10.0 /. 2.0 }
+      { amt = 10.; total_cost = 2.0; unit_price = 10.0 /. 2.0 }
       (* lemon = [ { amt = 10; total_cost = 2.0; unit_price = 10.0 /. 2.0 }; {
          amt = 20; total_cost = 3.75; unit_price = 20.0 /. 3.75 }; ]; cups = [ {
          amt = 10; total_cost = 2.0; unit_price = 10.0 /. 2.0 }; { amt = 20;
@@ -46,7 +46,7 @@ let get_cup_total_cost = purchase_options.cup.total_cost
 let get_cup_amt = purchase_options.cup.amt
 let get_sugar_total_cost = purchase_options.sugar.total_cost
 let get_sugar_amt = purchase_options.sugar.amt
-let init_cup = { lemon = 0; sugar = 0; water = 0 }
+let init_cup = { lemon = 0.; sugar = 0.; water = 0. }
 
 let get_cup =
   { lemon = init_cup.lemon; sugar = init_cup.sugar; water = init_cup.water }
